+++ src/main/java/com/devnity/devnity/domain/mapgakco/service/mapgakco/MapgakcoService.java
@@ -1,14 +1,11 @@
-package com.devnity.devnity.domain.mapgakco.service;
+package com.devnity.devnity.domain.mapgakco.service.mapgakco;
 
-import com.devnity.devnity.common.error.exception.EntityNotFoundException;
-import com.devnity.devnity.common.error.exception.ErrorCode;
 import com.devnity.devnity.domain.mapgakco.converter.MapgakcoConverter;
 import com.devnity.devnity.domain.mapgakco.dto.mapgakco.request.MapgakcoCreateRequest;
-import com.devnity.devnity.domain.mapgakco.entity.Mapgakco;
 import com.devnity.devnity.domain.mapgakco.entity.MapgakcoStatus;
 import com.devnity.devnity.domain.mapgakco.repository.MapgakcoRepository;
+import com.devnity.devnity.domain.mapgakco.service.MapgakcoFacadeService;
 import com.devnity.devnity.domain.user.entity.User;
-import com.devnity.devnity.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,22 +17,17 @@ public class MapgakcoService {
 
   private final MapgakcoConverter mapgakcoConverter;
   private final MapgakcoRepository mapgakcoRepository;
-  private final UserRepository userRepository;
+  private final MapgakcoFacadeService mapgakcoFacadeService;
 
   @Transactional
   public MapgakcoStatus create(Long userId, MapgakcoCreateRequest request) {
-    // Todo : 중복코드 리펙토링 -> facade 클래스
-    User user = userRepository.findById(userId)
-      .orElseThrow(() -> new EntityNotFoundException("User is not found."));
+    User user = mapgakcoFacadeService.findUserById(userId);
     return mapgakcoRepository.save(mapgakcoConverter.toMapgakco(user, request)).getStatus();
   }
 
   @Transactional
-  public MapgakcoStatus delete(Long mapgakcoId) {
-    Mapgakco mapgakco = mapgakcoRepository.findById(mapgakcoId)
-      .orElseThrow(
-        () -> new EntityNotFoundException("Mapgakco is not found.", ErrorCode.MAPGAKCO_NOT_FOUND));
-    return mapgakco.updateStatus(MapgakcoStatus.DELETED);
+  public void delete(Long mapgakcoId) {
+    mapgakcoFacadeService.findMapgakcoById(mapgakcoId).delete();
   }
 
 }
